Move _crawl_events from watch_contract.
